my-react-app/src/backend/requirements.txt
# Flask y extensiones principales
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1

# Base de datos
SQLAlchemy==2.0.23

# Seguridad
bcrypt==4.1.2
python-dotenv==1.0.0

# APIs de terceros
sendgrid==6.11.0
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
 
# Utilidades
python-dateutil==2.8.2
pytz==2023.3

# Servidor de producción
gunicorn==21.2.0

# Testing (opcional)
pytest==7.4.3
pytest-flask==1.3.0
